requirements.txt
automata-lib
graphviz